requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
pillow==11.3.0
PyJWT==2.9.0
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
